Location and Region values are now floats.

+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/positioning/ElementPositionMemento.java
@@ -17,11 +17,11 @@ public class ElementPositionMemento extends PositionMemento {
         position = new Location(l);
     }
 
-    public int getX() {
+    public float getX() {
         return position.getX();
     }
 
-    public int getY() {
+    public float getY() {
         return position.getY();
     }
 }
